+++ tests/test_openapi_v3.py
@@ -2188,3 +2188,346 @@ components:
                     - 3
 """.strip()
     )
+
+
+@pytest.mark.asyncio
+async def test_mount_oad_generation(serializer: Serializer):
+    """
+    Tests support for OAD generation of mounted apps, using the options:
+
+    parent.mount_registry.auto_events = True
+    parent.mount_registry.handle_docs = True
+    """
+    parent = Application(show_error_details=True)
+    parent.mount_registry.auto_events = True
+    parent.mount_registry.handle_docs = True
+
+    docs = OpenAPIHandler(info=Info(title="Parent API", version="0.0.1"))
+    docs.bind_app(parent)
+
+    @dataclass
+    class CreateCatInput:
+        name: str
+        email: str
+        foo: int
+
+    @dataclass
+    class CreateDogInput:
+        name: str
+        email: str
+        example: int
+
+    @dataclass
+    class CreateParrotInput:
+        name: str
+        email: str
+
+    @parent.router.get("/")
+    def a_home():
+        """Parent root."""
+        return "Hello, from the parent app - for information, navigate to /docs"
+
+    @parent.router.get("/cats")
+    def get_cats_conflicting():
+        """Conflict! This will be overridden by the child app route!"""
+        return "CONFLICT"
+
+    child_1 = Application()
+
+    @child_1.router.get("/")
+    def get_cats():
+        """Gets a list of cats."""
+        return "Gets a list of cats."
+
+    @child_1.router.post("/")
+    def create_cat(data: CreateCatInput):
+        """Creates a new cat."""
+        return "Creates a new cat."
+
+    @child_1.router.delete("/{cat_id}")
+    def delete_cat(cat_id: str):
+        """Deletes a cat by id."""
+        return "Deletes a cat by id."
+
+    child_2 = Application()
+
+    @child_2.router.get("/")
+    def get_dogs():
+        """Gets a list of dogs."""
+        return "Gets a list of dogs."
+
+    @child_2.router.post("/")
+    def create_dog(data: CreateDogInput):
+        """Creates a new dog."""
+        return "Creates a new dog."
+
+    @child_2.router.delete("/{dog_id}")
+    def delete_dog(dog_id: str):
+        """Deletes a dog by id."""
+        return "Deletes a dog by id."
+
+    child_3 = Application()
+
+    @child_3.router.get("/")
+    def get_parrots():
+        """Gets a list of parrots."""
+        return "Gets a list of parrots"
+
+    @child_3.router.post("/")
+    def create_parrot(data: CreateParrotInput):
+        """Creates a new parrot."""
+        return "Creates a new parrot"
+
+    @child_3.router.delete("/{parrot_id}")
+    def delete_parrot(parrot_id: str):
+        """Deletes a parrot by id."""
+        return "Deletes a parrot by id."
+
+    parent.mount("/cats", child_1)
+    parent.mount("/dogs", child_2)
+    parent.mount("/parrots", child_3)
+
+    await parent.start()
+
+    yaml = serializer.to_yaml(docs.generate_documentation(parent))
+
+    assert (
+        yaml.strip()
+        == """
+openapi: 3.0.3
+info:
+    title: Parent API
+    version: 0.0.1
+paths:
+    /:
+        get:
+            responses: {}
+            operationId: a_home
+            summary: Parent root.
+            description: Parent root.
+    /cats:
+        get:
+            responses: {}
+            operationId: get_cats
+            summary: Gets a list of cats.
+            description: Gets a list of cats.
+        post:
+            responses: {}
+            operationId: create_cat
+            summary: Creates a new cat.
+            description: Creates a new cat.
+            parameters: []
+            requestBody:
+                content:
+                    application/json:
+                        schema:
+                            $ref: '#/components/schemas/CreateCatInput'
+                required: true
+    /cats/{cat_id}:
+        delete:
+            responses: {}
+            operationId: delete_cat
+            summary: Deletes a cat by id.
+            description: Deletes a cat by id.
+            parameters:
+            -   name: cat_id
+                in: path
+                schema:
+                    type: string
+                    nullable: false
+                description: ''
+                required: true
+    /dogs:
+        get:
+            responses: {}
+            operationId: get_dogs
+            summary: Gets a list of dogs.
+            description: Gets a list of dogs.
+        post:
+            responses: {}
+            operationId: create_dog
+            summary: Creates a new dog.
+            description: Creates a new dog.
+            parameters: []
+            requestBody:
+                content:
+                    application/json:
+                        schema:
+                            $ref: '#/components/schemas/CreateDogInput'
+                required: true
+    /dogs/{dog_id}:
+        delete:
+            responses: {}
+            operationId: delete_dog
+            summary: Deletes a dog by id.
+            description: Deletes a dog by id.
+            parameters:
+            -   name: dog_id
+                in: path
+                schema:
+                    type: string
+                    nullable: false
+                description: ''
+                required: true
+    /parrots:
+        get:
+            responses: {}
+            operationId: get_parrots
+            summary: Gets a list of parrots.
+            description: Gets a list of parrots.
+        post:
+            responses: {}
+            operationId: create_parrot
+            summary: Creates a new parrot.
+            description: Creates a new parrot.
+            parameters: []
+            requestBody:
+                content:
+                    application/json:
+                        schema:
+                            $ref: '#/components/schemas/CreateParrotInput'
+                required: true
+    /parrots/{parrot_id}:
+        delete:
+            responses: {}
+            operationId: delete_parrot
+            summary: Deletes a parrot by id.
+            description: Deletes a parrot by id.
+            parameters:
+            -   name: parrot_id
+                in: path
+                schema:
+                    type: string
+                    nullable: false
+                description: ''
+                required: true
+components:
+    schemas:
+        CreateCatInput:
+            type: object
+            required:
+            - name
+            - email
+            - foo
+            properties:
+                name:
+                    type: string
+                    nullable: false
+                email:
+                    type: string
+                    nullable: false
+                foo:
+                    type: integer
+                    format: int64
+                    nullable: false
+        CreateDogInput:
+            type: object
+            required:
+            - name
+            - email
+            - example
+            properties:
+                name:
+                    type: string
+                    nullable: false
+                email:
+                    type: string
+                    nullable: false
+                example:
+                    type: integer
+                    format: int64
+                    nullable: false
+        CreateParrotInput:
+            type: object
+            required:
+            - name
+            - email
+            properties:
+                name:
+                    type: string
+                    nullable: false
+                email:
+                    type: string
+                    nullable: false
+""".strip()
+    )
+
+
+@pytest.mark.asyncio
+async def test_mount_oad_generation_sub_children(serializer: Serializer):
+    """
+    Tests support for OAD generation of mounted apps, using the options:
+
+    parent.mount_registry.auto_events = True
+    parent.mount_registry.handle_docs = True
+    """
+    parent = Application(show_error_details=True)
+    parent.mount_registry.auto_events = True
+    parent.mount_registry.handle_docs = True
+
+    docs = OpenAPIHandler(info=Info(title="Parent API", version="0.0.1"))
+    docs.bind_app(parent)
+
+    @parent.router.get("/")
+    def a_home():
+        """Parent root."""
+        return "Hello, from the parent app - for information, navigate to /docs"
+
+    child_1 = Application()
+    child_2 = Application()
+    child_3 = Application()
+
+    @child_1.router.get("/")
+    def child_1_home():
+        return "Child 1 home."
+
+    @child_2.router.get("/")
+    def child_2_home():
+        return "Child 2 home."
+
+    @child_3.router.get("/")
+    def child_3_home():
+        return "Child 3 home."
+
+    child_1.mount_registry.auto_events = True
+    child_1.mount_registry.handle_docs = True
+    child_2.mount_registry.auto_events = True
+    child_2.mount_registry.handle_docs = True
+
+    parent.mount("/child-1", child_1)
+    child_1.mount("/child-2", child_2)
+    child_2.mount("/child-3", child_3)
+
+    await parent.start()
+
+    yaml = serializer.to_yaml(docs.generate_documentation(parent))
+
+    assert (
+        yaml.strip()
+        == """
+openapi: 3.0.3
+info:
+    title: Parent API
+    version: 0.0.1
+paths:
+    /:
+        get:
+            responses: {}
+            operationId: a_home
+            summary: Parent root.
+            description: Parent root.
+    /child-1:
+        get:
+            responses: {}
+            operationId: child_1_home
+    /child-1/child-2:
+        get:
+            responses: {}
+            operationId: child_2_home
+    /child-1/child-2/child-3:
+        get:
+            responses: {}
+            operationId: child_3_home
+components: {}
+""".strip()
+    )