+++ CHANGELOG.md
@@ -5,12 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.5] - 2022-03-??
+## [1.2.5] - 2022-03-12 :dove:
 * Improves WebSocket to handle built-in exception types: Unauthorized, HTTPException
 * Adds built-in support for [Anti Forgery validation](https://www.neoteroi.dev/blacksheep/anti-request-forgery) to protect against Cross-Site Request Forgery (XSRF/CSRF) attacks
 * Modifies the Request and Response classes to support weak references
 * Adds the possibility to use `**kwargs` in view functions, returning HTML built
   using Jinja2
+* Adds support for automatic handling of child application events when BlackSheep
+  applications are mounted into a parent BlackSheep application
+* Adds support for OpenAPI Documentation generated for children BlackSheep apps,
+  when using mounts
+* Corrects bugs that prevented mounted routes to work recursively in descendants
+* Updates dependencies
 
 ## [1.2.4] - 2022-02-13 :cat:
 * Modifies the `WebSocket` class to support built-in binders
@@ -61,7 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrades `Jinja2` dependency to version `3.0.2`
 - Modifies `setup.py` dependencies to be less strict (`~=` instead of `==`)
 
-## [1.0.9] - 2021-07-14 :italy:
+## [1.0.9] - 2021-07-14 🇮🇹
 - Adds support for application mounts (see [discussion #160](https://github.com/Neoteroi/BlackSheep/discussions/160))
 - Applies sorting of imports using `isort`, enforces linters in the CI pipeline
   with both `black` and `isort`